normal fix

+++ config/stage_1_style_tests/check_lint.py
@@ -2,6 +2,7 @@
 Check lint for code style in Python code.
 """
 # pylint: disable=duplicate-code
+import argparse
 import shutil
 import subprocess
 from os import listdir
@@ -61,11 +62,23 @@ def check_lint_level(lint_output: bytes, target_score: int) -> subprocess.Comple
     ]
     return _run_console_tool(str(choose_python_exe()), lint_level_args, debug=True)
 
+def parse_arguments() -> argparse.Namespace:
+    """
+    Parse command line arguments.
+
+    Returns:
+        argparse.Namespace: Parsed arguments.
+    """
+    parser = argparse.ArgumentParser(description="Run check_lint.py checks for each lab.")
+    parser.add_argument("--repository_type", help="Type of the repository (public/private)")
+    return parser.parse_args()
 
 def main() -> None:
     """
     Run lint checks for the project.
     """
+    args = parse_arguments()
+    repository_type = args.repository_type
     project_config = ProjectConfig(PROJECT_CONFIG_PATH)
     labs_list = project_config.get_labs_paths()
 
@@ -99,12 +112,15 @@ def main() -> None:
 
     for lab_name in labs_list:
         lab_path = PROJECT_ROOT / lab_name
-        tests_subdir = lab_path / 'tests'
-        if tests_subdir.exists() and tests_subdir.is_dir():
-            shutil.rmtree(tests_subdir)
+        if repository_type == "public":
+            tests_subdir = lab_path / 'tests'
+            if tests_subdir.exists() and tests_subdir.is_dir():
+                shutil.rmtree(tests_subdir)
 
         if "settings.json" in listdir(lab_path):
             target_score = LabSettings(PROJECT_ROOT / f"{lab_path}/settings.json").target_score
+            if target_score == 0: 
+                continue
 
             print(f"Running lint for lab {lab_path}")
             completed_process = check_lint_on_paths(
